vm/bhyve: enable running as an unprivileged user
- Add a bootrom option which specifies the boot firmware to use.
  Typically this would be /usr/local/share/uefi-firmware/BHYVE_UEFI.fd
  on amd64.  The legacy bhyveload mechanism is only available to
  privileged users.
- Instead of using fancy ZFS tricks to avoid copying the VM image, just
  copy it into each instance's workdir: use cp(1) so that this
  automatically takes advantage of ZFS block cloning, which should
  transparently avoid much of the overhead of copying.

+++ vm/bhyve/bhyve.go
@@ -33,9 +33,10 @@ type Config struct {
 	Bridge  string `json:"bridge"`  // name of network bridge device, optional
 	Count   int    `json:"count"`   // number of VMs to use
 	CPU     int    `json:"cpu"`     // number of VM vCPU
-	HostIP  string `json:"hostip"`  // VM host IP address
+	HostIP  string `json:"hostip"`  // VM host IP address, optional
 	Mem     string `json:"mem"`     // amount of VM memory
-	Dataset string `json:"dataset"` // ZFS dataset containing VM image
+	Dataset string `json:"dataset"` // ZFS dataset containing VM image, optional
+	BootRom string `json:"bootrom"` // path to UEFI bootrom, optional
 }
 
 type Pool struct {
@@ -96,39 +97,51 @@ func (pool *Pool) Create(_ context.Context, workdir string, index int) (vmimpl.I
 		vmName: fmt.Sprintf("syzkaller-%v-%v", pool.env.Name, index),
 	}
 
-	dataset := inst.cfg.Dataset
-	mountpoint, err := osutil.RunCmd(time.Minute, "", "zfs", "get", "-H", "-o", "value", "mountpoint", dataset)
-	if err != nil {
-		return nil, err
-	}
+	if inst.cfg.Dataset != "" {
+		dataset := inst.cfg.Dataset
+		mountpoint, err := osutil.RunCmd(time.Minute, "", "zfs", "get", "-H", "-o", "value", "mountpoint", dataset)
+		if err != nil {
+			return nil, err
+		}
 
-	snapshot := fmt.Sprintf("%v@bhyve-%v", dataset, inst.vmName)
-	clone := fmt.Sprintf("%v/bhyve-%v", dataset, inst.vmName)
+		snapshot := fmt.Sprintf("%v@bhyve-%v", dataset, inst.vmName)
+		clone := fmt.Sprintf("%v/bhyve-%v", dataset, inst.vmName)
 
-	prefix := strings.TrimSuffix(string(mountpoint), "\n") + "/"
-	image := strings.TrimPrefix(pool.env.Image, prefix)
-	if image == pool.env.Image {
-		return nil, fmt.Errorf("image file %v not contained in dataset %v", image, prefix)
-	}
-	inst.image = prefix + fmt.Sprintf("bhyve-%v", inst.vmName) + "/" + image
+		prefix := strings.TrimSuffix(string(mountpoint), "\n") + "/"
+		image := strings.TrimPrefix(pool.env.Image, prefix)
+		if image == pool.env.Image {
+			return nil, fmt.Errorf("image file %v not contained in dataset %v", image, prefix)
+		}
+		inst.image = prefix + fmt.Sprintf("bhyve-%v", inst.vmName) + "/" + image
 
-	// Stop the instance from a previous run in case it's still running.
-	osutil.RunCmd(time.Minute, "", "bhyvectl", "--destroy", fmt.Sprintf("--vm=%v", inst.vmName))
-	// Destroy a lingering snapshot and clone.
-	osutil.RunCmd(time.Minute, "", "zfs", "destroy", "-R", snapshot)
+		// Stop the instance from a previous run in case it's still running.
+		osutil.RunCmd(time.Minute, "", "bhyvectl", "--destroy", fmt.Sprintf("--vm=%v", inst.vmName))
+		// Destroy a lingering snapshot and clone.
+		osutil.RunCmd(time.Minute, "", "zfs", "destroy", "-R", snapshot)
 
-	// Create a snapshot of the data set containing the VM image.
-	// bhyve will use a clone of the snapshot, which gets recreated every time the VM
-	// is restarted. This is all to work around bhyve's current lack of an
-	// image snapshot facility.
-	if _, err := osutil.RunCmd(time.Minute, "", "zfs", "snapshot", snapshot); err != nil {
-		inst.Close()
-		return nil, err
-	}
-	inst.snapshot = snapshot
-	if _, err := osutil.RunCmd(time.Minute, "", "zfs", "clone", snapshot, clone); err != nil {
-		inst.Close()
-		return nil, err
+		// Create a snapshot of the data set containing the VM image.
+		// bhyve will use a clone of the snapshot, which gets recreated every time the VM
+		// is restarted. This is all to work around bhyve's current lack of an
+		// image snapshot facility.
+		if _, err := osutil.RunCmd(time.Minute, "", "zfs", "snapshot", snapshot); err != nil {
+			inst.Close()
+			return nil, err
+		}
+		inst.snapshot = snapshot
+		if _, err := osutil.RunCmd(time.Minute, "", "zfs", "clone", snapshot, clone); err != nil {
+			inst.Close()
+			return nil, err
+		}
+	} else {
+		inst.image = filepath.Join(workdir, filepath.Base(pool.env.Image)+fmt.Sprintf("-%v", inst.vmName))
+		// Use cp(1) here to help ensure that the image file gets
+		// copied using copy_file_range(2), which will leverage block
+		// cloning on ZFS.
+		_, err := osutil.RunCmd(time.Minute, "", "cp", "-f", pool.env.Image, inst.image)
+		if err != nil {
+			inst.Close()
+			return nil, err
+		}
 	}
 
 	if inst.cfg.Bridge != "" {
@@ -153,21 +166,8 @@ func (pool *Pool) Create(_ context.Context, workdir string, index int) (vmimpl.I
 }
 
 func (inst *instance) Boot() error {
-	loaderArgs := []string{
-		"-c", "stdio",
-		"-m", inst.cfg.Mem,
-		"-d", inst.image,
-		"-e", "autoboot_delay=0",
-		inst.vmName,
-	}
-
 	// Stop the instance from the previous run in case it's still running.
 	osutil.RunCmd(time.Minute, "", "bhyvectl", "--destroy", fmt.Sprintf("--vm=%v", inst.vmName))
-
-	_, err := osutil.RunCmd(time.Minute, "", "bhyveload", loaderArgs...)
-	if err != nil {
-		return err
-	}
 
 	netdev := ""
 	if inst.tapdev != "" {
@@ -188,6 +188,22 @@ func (inst *instance) Boot() error {
 		"-s", fmt.Sprintf("3:0,virtio-blk,%v", inst.image),
 		"-l", "com1,stdio",
 		inst.vmName,
+	}
+
+	if inst.cfg.BootRom == "" {
+		loaderArgs := []string{
+			"-c", "stdio",
+			"-m", inst.cfg.Mem,
+			"-d", inst.image,
+			"-e", "autoboot_delay=0",
+			inst.vmName,
+		}
+		_, err := osutil.RunCmd(time.Minute, "", "bhyveload", loaderArgs...)
+		if err != nil {
+			return err
+		}
+	} else {
+		bhyveArgs = append([]string{"-M", "-l", fmt.Sprintf("bootrom,%v", inst.cfg.BootRom)}, bhyveArgs...)
 	}
 
 	outr, outw, err := osutil.LongPipe()
